backend_python/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pillow==10.0.1
scikit-image==0.23.2
scikit-learn==1.5.0
imbalanced-learn==0.12.3
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
pathspec==0.12.1
filelock==3.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
